Add dockerfile and compose to test the interval functionality using environemnt.
Adding unittest

improving the code and remove the warnings
