--- mzu.libs.rollercoaster.examples.console/Program.cs
@@ -1,10 +1,0 @@
-﻿using mzu.libs.rollercoaster;
-using Mzu.Libs.Rollercoaster;
-
-RollerCoasterMethodExecutor.ExecuteRollerCoasterMethods();
-
-Console.WriteLine("Press any key to stop...");
-Console.ReadKey();
-
-
-
