--- mzu.libs.rollercoaster.examples.console/mzu.libs.rollercoaster.examples.console.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\mzu.libs.rollercoaster\Mzu.Libs.Rollercoaster.csproj" />
-  </ItemGroup>
-
-</Project>
