--- mzu.libs.rollercoaster.examples.webapi/Repository/IExampleRepository.cs
@@ -1,7 +1,0 @@
-﻿namespace mzu.libs.rollercoaster.examples.webapi.Repository;
-
-public interface IExampleRepository
-{
-    Task<List<string>> GetData();
-    Task SetData(string data);
-}
