--- mzu.libs.rollercoaster.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33122.133
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mzu.Libs.Rollercoaster", "mzu.libs.rollercoaster\Mzu.Libs.Rollercoaster.csproj", "{D14F355F-73BB-4879-B08A-80D903B4D072}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "mzu.libs.rollercoaster.examples.console", "mzu.libs.rollercoaster.examples.console\mzu.libs.rollercoaster.examples.console.csproj", "{E4117007-1651-4E8A-8F78-5562CF3E034E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "mzu.libs.rollercoaster.examples.webapi", "mzu.libs.rollercoaster.examples.webapi\mzu.libs.rollercoaster.examples.webapi.csproj", "{30205C11-92FB-4B11-9635-586C91B2D89E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D14F355F-73BB-4879-B08A-80D903B4D072}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D14F355F-73BB-4879-B08A-80D903B4D072}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D14F355F-73BB-4879-B08A-80D903B4D072}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D14F355F-73BB-4879-B08A-80D903B4D072}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E4117007-1651-4E8A-8F78-5562CF3E034E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E4117007-1651-4E8A-8F78-5562CF3E034E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E4117007-1651-4E8A-8F78-5562CF3E034E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E4117007-1651-4E8A-8F78-5562CF3E034E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{30205C11-92FB-4B11-9635-586C91B2D89E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30205C11-92FB-4B11-9635-586C91B2D89E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30205C11-92FB-4B11-9635-586C91B2D89E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30205C11-92FB-4B11-9635-586C91B2D89E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {95E2B8A3-E515-4CC4-896B-36C7A62C3A8A}
-	EndGlobalSection
-EndGlobal
