+++ src/Mzu.Libs.Rollercoaster.Unittest/Exceptions/ObjectLockException.cs
@@ -1,0 +1,5 @@
+﻿namespace Mzu.Libs.Rollercoaster.Unittest.Exceptions;
+
+internal class ObjectLockException: Exception
+{
+}
