+++ src/mzu.libs.rollercoaster.examples.webapi/Helpers/StringGenerator.cs
