+++ src/mzu.libs.rollercoaster.examples.webapi/Repository/IExampleRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace mzu.libs.rollercoaster.examples.webapi.Repository;
+
+public interface IExampleRepository
+{
+    List<string> GetData();
+    void SetData(string data);
+}
