+++ src/mzu.libs.rollercoaster.examples.webapi/appsettings.Development.json
