+++ src/mzu.libs.rollercoaster.examples.webapi/appsettings.json
