+++ src/mzu.libs.rollercoaster.examples.webapi/launchSettings.json
@@ -1,0 +1,11 @@
+{
+  "profiles": {
+    "Docker Compose": {
+      "commandName": "DockerCompose",
+      "commandVersion": "1.0",
+      "serviceActions": {
+        "mzu.libs.rollercoaster.examples.webapi": "StartDebugging"
+      }
+    }
+  }
+}
