+++ src/mzu.libs.rollercoaster/Configurations/Session.cs
