+++ src/mzu.libs.rollercoaster/Mzu.Libs.Rollercoaster.csproj
@@ -1,0 +1,41 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <Title>RollerCoaster a lightweight method Scheduler library</Title>
+    <Authors>$(AssemblyName), Mohamad Ravaei</Authors>
+    <Description>The RollerCoaster Project is a lightweight library that allows you to schedule and run methods at specific intervals using background services. It provides an easy way to execute methods periodically without blocking the main thread.</Description>
+    <Copyright>MIT License</Copyright>
+	<PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageProjectUrl>https://github.com/ravaee/Mzu.Libs.Rollercoaster/tree/master</PackageProjectUrl>
+    <PackageIcon>rollercoaster-831513_960_720.jpg</PackageIcon>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <RepositoryUrl>https://github.com/ravaee/Mzu.Libs.Rollercoaster/tree/master</RepositoryUrl>
+    <PackageTags>Job, Scheduler, Cronjob</PackageTags>
+    <PackageReleaseNotes>- Adding tests
+- Extend the Readme file</PackageReleaseNotes>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Exceptions\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="..\README.md">
+      <Pack>True</Pack>
+      <PackagePath>\</PackagePath>
+    </None>
+    <None Include="..\rollercoaster-831513_960_720.jpg">
+      <Pack>True</Pack>
+      <PackagePath>\</PackagePath>
+    </None>
+  </ItemGroup>
+
+</Project>
