+++ src/mzu.libs.rollercoaster/Options/RollerCoasterOptions.cs
